alternative HTML which doesn't use data-main
This is particularly useful when using the mainConfigFile option in your build config in order to avoid duplicating your paths, shims, and deps in your require config and build config. When using the mainConfigFile option, the require configuration needs to be specified like "require.config({...})" instead of "require = {...}" therefore the require.js script needs to be loaded first.